#5352: Define interface for HealthGroupsCache
